Combine copyright handlers into single file
